requirements.txt
pandas
numpy
networkx
scikit-learn
openai
tqdm
python-dotenvpython
pyarrow